danjhd/v3 refactor utils (#13)
* dedicated api related code layer

* update files to new generation script versions

* refactored utils.py into separate files
