+++ functions/api_sources/app.py
@@ -2,8 +2,6 @@ import json
 import os
 from http import HTTPStatus
 
-# pylint: disable=no-member
-import constants
 from aws_lambda_powertools import Logger, Metrics, Tracer
 from aws_lambda_powertools.event_handler import (
     APIGatewayRestResolver,
@@ -18,19 +16,22 @@ from aws_lambda_powertools.event_handler.exceptions import (
 from aws_lambda_powertools.event_handler.openapi.params import Path
 from aws_lambda_powertools.logging import correlation_paths
 from aws_lambda_powertools.utilities.typing import LambdaContext
-from schema import Source, Tags
-from typing_extensions import Annotated
-from utils import (
+from neptune import (
     check_node_exists,
-    generate_link_url,
-    get_username,
-    model_dump,
-    publish_event,
+    enhance_resources,
     query_node,
     query_node_property,
     query_node_tags,
     query_sources,
     set_node_property,
+)
+from schema import Source, Tags
+from typing_extensions import Annotated
+from utils import (
+    generate_link_url,
+    get_username,
+    model_dump,
+    publish_event,
     validate_query_string,
 )
 
@@ -41,6 +42,8 @@ metrics = Metrics(namespace="Powertools")
 
 record_type = "source"
 event_bus = os.environ["EVENT_BUS"]
+
+SOURCE_ID_PATTERN = Source.model_fields["id"].metadata[0].pattern
 
 
 @app.head("/sources")
@@ -75,9 +78,7 @@ def list_sources():
 @app.head("/sources/<sourceId>")
 @app.get("/sources/<sourceId>")
 @tracer.capture_method(capture_response=False)
-def get_source_details(
-    sourceId: Annotated[str, Path(pattern=constants.SOURCE_ID_PATTERN)]
-):
+def get_source_details(sourceId: Annotated[str, Path(pattern=SOURCE_ID_PATTERN)]):
     try:
         item = query_node(record_type, sourceId)
     except ValueError as e:
@@ -90,9 +91,7 @@ def get_source_details(
 @app.head("/sources/<sourceId>/tags")
 @app.get("/sources/<sourceId>/tags")
 @tracer.capture_method(capture_response=False)
-def get_source_tags(
-    sourceId: Annotated[str, Path(pattern=constants.SOURCE_ID_PATTERN)]
-):
+def get_source_tags(sourceId: Annotated[str, Path(pattern=SOURCE_ID_PATTERN)]):
     try:
         tags = query_node_tags(record_type, sourceId)
     except ValueError as e:
@@ -106,7 +105,7 @@ def get_source_tags(
 @app.get("/sources/<sourceId>/tags/<name>")
 @tracer.capture_method(capture_response=False)
 def get_source_tag_value(
-    sourceId: Annotated[str, Path(pattern=constants.SOURCE_ID_PATTERN)], name: str
+    sourceId: Annotated[str, Path(pattern=SOURCE_ID_PATTERN)], name: str
 ):
     try:
         tags = query_node_tags(record_type, sourceId)
@@ -122,7 +121,7 @@ def get_source_tag_value(
 @app.put("/sources/<sourceId>/tags/<name>")
 @tracer.capture_method(capture_response=False)
 def put_source_tag_value(
-    sourceId: Annotated[str, Path(pattern=constants.SOURCE_ID_PATTERN)], name: str
+    sourceId: Annotated[str, Path(pattern=SOURCE_ID_PATTERN)], name: str
 ):
     if not check_node_exists(record_type, sourceId):
         raise NotFoundError(
@@ -147,7 +146,7 @@ def put_source_tag_value(
 @app.delete("/sources/<sourceId>/tags/<name>")
 @tracer.capture_method(capture_response=False)
 def delete_source_tag(
-    sourceId: Annotated[str, Path(pattern=constants.SOURCE_ID_PATTERN)], name: str
+    sourceId: Annotated[str, Path(pattern=SOURCE_ID_PATTERN)], name: str
 ):
     try:
         item = query_node(record_type, sourceId)
@@ -172,9 +171,7 @@ def delete_source_tag(
 @app.head("/sources/<sourceId>/description")
 @app.get("/sources/<sourceId>/description")
 @tracer.capture_method(capture_response=False)
-def get_source_description(
-    sourceId: Annotated[str, Path(pattern=constants.SOURCE_ID_PATTERN)]
-):
+def get_source_description(sourceId: Annotated[str, Path(pattern=SOURCE_ID_PATTERN)]):
     try:
         description = query_node_property(record_type, sourceId, "description")
     except ValueError as e:
@@ -186,9 +183,7 @@ def get_source_description(
 
 @app.put("/sources/<sourceId>/description")
 @tracer.capture_method(capture_response=False)
-def put_source_description(
-    sourceId: Annotated[str, Path(pattern=constants.SOURCE_ID_PATTERN)]
-):
+def put_source_description(sourceId: Annotated[str, Path(pattern=SOURCE_ID_PATTERN)]):
     if not check_node_exists(record_type, sourceId):
         raise NotFoundError("The requested Source does not exist.")  # 404
     try:
@@ -212,7 +207,7 @@ def put_source_description(
 @app.delete("/sources/<sourceId>/description")
 @tracer.capture_method(capture_response=False)
 def delete_source_description(
-    sourceId: Annotated[str, Path(pattern=constants.SOURCE_ID_PATTERN)]
+    sourceId: Annotated[str, Path(pattern=SOURCE_ID_PATTERN)]
 ):
     if not check_node_exists(record_type, sourceId):
         raise NotFoundError("The Source ID in the path is invalid.")  # 404
@@ -231,9 +226,7 @@ def delete_source_description(
 @app.head("/sources/<sourceId>/label")
 @app.get("/sources/<sourceId>/label")
 @tracer.capture_method(capture_response=False)
-def get_source_label(
-    sourceId: Annotated[str, Path(pattern=constants.SOURCE_ID_PATTERN)]
-):
+def get_source_label(sourceId: Annotated[str, Path(pattern=SOURCE_ID_PATTERN)]):
     try:
         label = query_node_property(record_type, sourceId, "label")
     except ValueError as e:
@@ -247,9 +240,7 @@ def get_source_label(
 
 @app.put("/sources/<sourceId>/label")
 @tracer.capture_method(capture_response=False)
-def put_source_label(
-    sourceId: Annotated[str, Path(pattern=constants.SOURCE_ID_PATTERN)]
-):
+def put_source_label(sourceId: Annotated[str, Path(pattern=SOURCE_ID_PATTERN)]):
     if not check_node_exists(record_type, sourceId):
         raise NotFoundError("The requested Source does not exist.")  # 404
     try:
@@ -272,9 +263,7 @@ def put_source_label(
 
 @app.delete("/sources/<sourceId>/label")
 @tracer.capture_method(capture_response=False)
-def delete_source_label(
-    sourceId: Annotated[str, Path(pattern=constants.SOURCE_ID_PATTERN)]
-):
+def delete_source_label(sourceId: Annotated[str, Path(pattern=SOURCE_ID_PATTERN)]):
     if not check_node_exists(record_type, sourceId):
         raise NotFoundError("The requested Source ID in the path is invalid.")  # 404
     username = get_username(app.current_event.request_context)
@@ -301,7 +290,12 @@ def lambda_handler(event: dict, context: LambdaContext) -> dict:
 @tracer.capture_method(capture_response=False)
 def get_event_resources(obj: dict) -> list:
     """Generate a list of event resources for the given source object."""
-    return [
-        f'tams:source:{obj["id"]}',
-        *[f"tams:source-collected-by:{c_id}" for c_id in obj.get("collected_by", [])],
-    ]
+    return enhance_resources(
+        [
+            f'tams:source:{obj["id"]}',
+            *[
+                f"tams:source-collected-by:{c_id}"
+                for c_id in obj.get("collected_by", [])
+            ],
+        ]
+    )
