+++ layers/utils/params.py
@@ -2,15 +2,15 @@ query_params = {
     "/sources": {
         "HEAD": {
             "label": "str",
-            "tag.name": "str",
-            "tag_exists.name": "bool",
+            "tag.{name}": "str",
+            "tag_exists.{name}": "bool",
             "page": "str",
             "limit": "int",
         },
         "GET": {
             "label": "str",
-            "tag.name": "str",
-            "tag_exists.name": "bool",
+            "tag.{name}": "str",
+            "tag_exists.{name}": "bool",
             "format": "str",
             "page": "str",
             "limit": "int",
@@ -23,8 +23,8 @@ query_params = {
             "format": "str",
             "codec": "str",
             "label": "str",
-            "tag.name": "str",
-            "tag_exists.name": "bool",
+            "tag.{name}": "str",
+            "tag_exists.{name}": "bool",
             "frame_width": "int",
             "frame_height": "int",
             "page": "str",
@@ -36,8 +36,8 @@ query_params = {
             "format": "str",
             "codec": "str",
             "label": "str",
-            "tag.name": "str",
-            "tag_exists.name": "bool",
+            "tag.{name}": "str",
+            "tag_exists.{name}": "bool",
             "frame_width": "int",
             "frame_height": "int",
             "page": "str",
