+++ layers/utils/schema.py
@@ -1,6 +1,6 @@
 # generated by datamodel-codegen:
 #   filename:  openapi.yaml
-#   timestamp: 2025-02-07T18:06:46+00:00
+#   timestamp: 2025-02-09T14:41:09+00:00
 
 from __future__ import annotations
 
@@ -32,6 +32,10 @@ class Uuid(
 
 
 class Contentformat(Enum):
+    """
+    Identifies the content format for a flow or source using a URN string.
+    """
+
     urn_x_nmos_format_video = "urn:x-nmos:format:video"
     urn_x_nmos_format_audio = "urn:x-nmos:format:audio"
     urn_x_nmos_format_data = "urn:x-nmos:format:data"
@@ -45,6 +49,10 @@ class Mimetype(RootModel[constr(pattern=r".*/.*")]):
 
 
 class Status(Enum):
+    """
+    Status of the delete request
+    """
+
     created = "created"
     started = "started"
     done = "done"
@@ -52,6 +60,10 @@ class Status(Enum):
 
 
 class Error(BaseModel):
+    """
+    Provides more information for the error status.
+    """
+
     type: str = Field(..., description="The error type name.")
     summary: str = Field(
         ..., description="Summary description of the error and causes."
@@ -79,6 +91,10 @@ class GetUrl(BaseModel):
 
 
 class Flowstoragepost(BaseModel):
+    """
+    Post data for the flow storage endpoint
+    """
+
     limit: Optional[int] = Field(
         None,
         description="Limit the number of storage segments in each response page. Implementations may specify their own default and maximum for the limit",
@@ -90,6 +106,10 @@ class Action(Enum):
 
 
 class Servicepost(BaseModel):
+    """
+    Post update to the service info
+    """
+
     name: Optional[str] = Field(None, description="The service instance name")
     description: Optional[str] = Field(
         None, description="The service instance description"
@@ -97,10 +117,18 @@ class Servicepost(BaseModel):
 
 
 class Type(Enum):
+    """
+    The type of the media store. This determines the endpoints for reading and writing media
+    """
+
     http_object_store = "http_object_store"
 
 
 class MediaStore(BaseModel):
+    """
+    Provide information about the media store for this service
+    """
+
     type: Type = Field(
         ...,
         description="The type of the media store. This determines the endpoints for reading and writing media",
@@ -108,6 +136,10 @@ class MediaStore(BaseModel):
 
 
 class Format(Enum):
+    """
+    The primary content type URN for the Source.
+    """
+
     urn_x_nmos_format_video = "urn:x-nmos:format:video"
     urn_x_nmos_format_audio = "urn:x-nmos:format:audio"
     urn_x_nmos_format_data = "urn:x-nmos:format:data"
@@ -115,6 +147,10 @@ class Format(Enum):
 
 
 class Tags(RootModel[Optional[Dict[str, str]]]):
+    """
+    Key value is a freeform string.
+    """
+
     root: Optional[Dict[str, str]] = None
 
 
@@ -129,6 +165,10 @@ class Timerange(
 
 
 class Webhookpost(BaseModel):
+    """
+    Register to receive updates via webhook
+    """
+
     url: str = Field(
         ...,
         description="The URL to which the API should make HTTP POST requests with event data",
@@ -149,7 +189,7 @@ class Webhookpost(BaseModel):
         ]
     ] = Field(
         None,
-        description="Limit Flow and Flow Segment events to Flows in the given list of Flow IDs",
+        description="Limit Flow related events to the given list of Flow IDs. Non-Flow related events are not affected",
     )
     source_ids: Optional[
         List[
@@ -159,35 +199,15 @@ class Webhookpost(BaseModel):
         ]
     ] = Field(
         None,
-        description="Limit Flow, Flow Segment and Source events to Sources in the given list of Source IDs",
-    )
-    flow_collected_by_ids: Optional[
-        List[
-            constr(
-                pattern=r"^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
-            )
-        ]
-    ] = Field(
-        None,
-        description="Limit Flow and Flow Segment events to those with Flow that is collected by a Flow Collection in the given list of Flow Collection IDs",
-    )
-    source_collected_by_ids: Optional[
-        List[
-            constr(
-                pattern=r"^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
-            )
-        ]
-    ] = Field(
-        None,
-        description="Limit Flow, Flow Segment and Source events to those with Source that is collected by a Source Collection in the given list of Source Collection IDs",
-    )
-    accept_get_urls: Optional[List[str]] = Field(
-        None,
-        description="List of labels of URLs to include in the `get_urls` property in `flows/segments_added` events. This option is the same as the `accept_get_urls` query parameter for the /flows/{flowId}/segments API endpoint, except that the labels are represented using a JSON array rather than a (comma separated list) string.",
+        description="Limit Source related events to the given list of Source IDs. Non-Source related events are not affected",
     )
 
 
 class Webhook(BaseModel):
+    """
+    Register to receive updates via webhook
+    """
+
     url: str = Field(
         ...,
         description="The URL to which the API should make HTTP POST requests with event data",
@@ -207,7 +227,7 @@ class Webhook(BaseModel):
         ]
     ] = Field(
         None,
-        description="Limit Flow and Flow Segment events to Flows in the given list of Flow IDs",
+        description="Limit Flow related events to the given list of Flow IDs. Non-Flow related events are not affected",
     )
     source_ids: Optional[
         List[
@@ -217,35 +237,15 @@ class Webhook(BaseModel):
         ]
     ] = Field(
         None,
-        description="Limit Flow, Flow Segment and Source events to Sources in the given list of Source IDs",
-    )
-    flow_collected_by_ids: Optional[
-        List[
-            constr(
-                pattern=r"^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
-            )
-        ]
-    ] = Field(
-        None,
-        description="Limit Flow and Flow Segment events to those with Flow that is collected by a Flow Collection in the given list of Flow Collection IDs",
-    )
-    source_collected_by_ids: Optional[
-        List[
-            constr(
-                pattern=r"^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
-            )
-        ]
-    ] = Field(
-        None,
-        description="Limit Flow, Flow Segment and Source events to those with Source that is collected by a Source Collection in the given list of Source Collection IDs",
-    )
-    accept_get_urls: Optional[List[str]] = Field(
-        None,
-        description="List of labels of URLs to include in the `get_urls` property in `flows/segments_added` events. This option is the same as the `accept_get_urls` query parameter for the /flows/{flowId}/segments API endpoint, except that the labels are represented using a JSON array rather than a (comma separated list) string.",
+        description="Limit Source related events to the given list of Source IDs. Non-Source related events are not affected",
     )
 
 
 class Eventstreamcommon(BaseModel):
+    """
+    Describes an event stream mechanism available in this implementation of TAMS
+    """
+
     name: str = Field(
         ...,
         description="Name of this type of event stream mechanism. Must be unique. Any name defined in this specification is reserved",
@@ -260,6 +260,10 @@ class Eventstreamcommon(BaseModel):
 
 
 class Format1(Enum):
+    """
+    The primary content type URN for the flow.
+    """
+
     urn_x_nmos_format_audio = "urn:x-nmos:format:audio"
 
 
@@ -274,6 +278,10 @@ class CodecParameters(BaseModel):
 
 
 class UncType(Enum):
+    """
+    The uncompressed audio multi-channel representation type. If codec is `audio/x-raw-int` or `audio/x-raw-float`, unc_type must be set.
+    """
+
     interleaved = "interleaved"
     planar = "planar"
     pairs = "pairs"
@@ -287,6 +295,10 @@ class UncParameters(BaseModel):
 
 
 class EssenceParameters(BaseModel):
+    """
+    Describes the parameters of the essence inside this audio Flow
+    """
+
     model_config = ConfigDict(
         extra="forbid",
     )
@@ -307,10 +319,18 @@ class EssenceParameters(BaseModel):
 
 
 class Format2(Enum):
+    """
+    The primary content type URN for the flow.
+    """
+
     urn_x_nmos_format_data = "urn:x-nmos:format:data"
 
 
 class EssenceParameters1(BaseModel):
+    """
+    Describes the parameters of the essence inside this data Flow
+    """
+
     model_config = ConfigDict(
         extra="forbid",
     )
@@ -321,19 +341,35 @@ class EssenceParameters1(BaseModel):
 
 
 class Format3(Enum):
+    """
+    The primary content type URN for the flow.
+    """
+
     urn_x_nmos_format_multi = "urn:x-nmos:format:multi"
 
 
 class Format4(Enum):
+    """
+    The primary content type URN for the flow.
+    """
+
     urn_x_nmos_format_video = "urn:x-nmos:format:video"
 
 
 class FrameRate(BaseModel):
+    """
+    The fixed number of frames per second. If this parameter is unset, the frame_rate is either unknown or variable.
+    """
+
     numerator: PositiveInt = Field(..., description="numerator")
     denominator: Optional[PositiveInt] = Field(1, description="denominator")
 
 
 class InterlaceMode(Enum):
+    """
+    Interlaced video mode for frames in this Flow
+    """
+
     progressive = "progressive"
     interlaced_tff = "interlaced_tff"
     interlaced_bff = "interlaced_bff"
@@ -341,6 +377,10 @@ class InterlaceMode(Enum):
 
 
 class Colorspace(Enum):
+    """
+    Colorspace used for the video
+    """
+
     BT601 = "BT601"
     BT709 = "BT709"
     BT2020 = "BT2020"
@@ -348,27 +388,47 @@ class Colorspace(Enum):
 
 
 class TransferCharacteristic(Enum):
+    """
+    Transfer characteristic
+    """
+
     SDR = "SDR"
     HLG = "HLG"
     PQ = "PQ"
 
 
 class AspectRatio(BaseModel):
+    """
+    The display aspect ratio. i.e. display_width / display_height
+    """
+
     numerator: PositiveInt = Field(..., description="numerator")
     denominator: PositiveInt = Field(..., description="denominator")
 
 
 class PixelAspectRatio(BaseModel):
+    """
+    The pixel aspect ratio. This is usually 1:1 (i.e. square pixels) for modern video. Some, usually older, video formats use non-square pixels e.g. some Standard Definition video. This is where that may be indicated.
+    """
+
     numerator: PositiveInt = Field(..., description="numerator")
     denominator: PositiveInt = Field(..., description="denominator")
 
 
 class ComponentType(Enum):
+    """
+    Picture component representation.
+    """
+
     YCbCr = "YCbCr"
     RGB = "RGB"
 
 
 class UncType1(Enum):
+    """
+    Uncompressed picture packing type. If codec is `video/raw`, unc_type must be set.
+    """
+
     planar = "planar"
     YUYV = "YUYV"
     UYVY = "UYVY"
@@ -410,6 +470,10 @@ class AvcParameters(BaseModel):
 
 
 class EssenceParameters2(BaseModel):
+    """
+    Describes the parameters of the essence inside this video Flow
+    """
+
     model_config = ConfigDict(
         extra="forbid",
     )
@@ -462,6 +526,10 @@ class EssenceParameters2(BaseModel):
 
 
 class Httprequest(BaseModel):
+    """
+    Gives information on a particular http request a client should perform
+    """
+
     model_config = ConfigDict(
         extra="allow",
     )
@@ -491,6 +559,10 @@ class ChannelNumber(RootModel[conint(ge=0)]):
 
 
 class AudioTrack(BaseModel):
+    """
+    Mapping for channels in audio tracks to the Flow channels
+    """
+
     channel_numbers: Optional[List[ChannelNumber]] = Field(
         None,
         description="Array of (zero-based) container channel numbers in Flow order",
@@ -502,12 +574,20 @@ class AudioTrack(BaseModel):
 
 
 class Mp2tsContainer(BaseModel):
+    """
+    Mapping to MPEG-2 Transport Stream containers, ISO/IEC 13818-1 or ITU-T Recommendation H.222.0
+    """
+
     pid: Optional[int] = Field(
         None, description="The packet ID for the elementary stream packets"
     )
 
 
 class MxfContainer(BaseModel):
+    """
+    Mapping to Material Exchange Format containers, SMPTE ST 377-1
+    """
+
     package_uid: Optional[
         constr(
             pattern=r"^urn:smpte:umid:[0-9a-fA-F]{8}(.[0-9a-fA-F]{8}){7}$|^urn:uuid:[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
@@ -517,10 +597,18 @@ class MxfContainer(BaseModel):
 
 
 class IsobmffContainer(BaseModel):
+    """
+    Mapping to ISO Base Media File Format (e.g. MP4 and MOV) containers, ISO/IEC 14496-12
+    """
+
     track_id: Optional[int] = Field(None, description="The track ID")
 
 
 class Containermapping(BaseModel):
+    """
+    Defines the location of Flow essence data in a container track
+    """
+
     track_index: Optional[conint(ge=0)] = Field(
         None,
         description="A zero-based and sequential track index in the container. This assumes a reliable ordering of tracks",
@@ -547,11 +635,19 @@ class Containermapping(BaseModel):
 
 
 class SegmentDuration(BaseModel):
+    """
+    The target flow segment duration in seconds. The duration for each segment may vary around this target value. See also the [Setting Flow Bit Rate Properties](https://github.com/bbc/tams/blob/main/docs/appnotes/0013-setting-flow-bit-rate-properties.md) AppNote for how this property can be used to calculate buffer sizes.
+    """
+
     numerator: PositiveInt = Field(..., description="numerator")
     denominator: Optional[PositiveInt] = Field(1, description="denominator")
 
 
 class Deletionrequest(BaseModel):
+    """
+    Describes an ongoing deletion request
+    """
+
     id: constr(
         pattern=r"^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
     ) = Field(..., description="Deletion Request ID")
@@ -594,6 +690,10 @@ class Deletionrequestslist(RootModel[List[Deletionrequest]]):
 
 
 class Flowsegment(BaseModel):
+    """
+    Provides the location and metadata of the media files corresponding to timerange segments of a Flow.
+    """
+
     object_id: str = Field(
         ..., description="The object store identifier for the media object."
     )
@@ -628,6 +728,10 @@ class Flowsegment(BaseModel):
 
 
 class PreItem(BaseModel):
+    """
+    An action
+    """
+
     action: Action
     bucket_id: Optional[str] = Field(
         None, description="The name of the bucket that needs to be created"
@@ -637,6 +741,10 @@ class PreItem(BaseModel):
 
 
 class MediaObject(BaseModel):
+    """
+    Information for a media object
+    """
+
     object_id: str = Field(
         ..., description="The object store identifier for the media object."
     )
@@ -645,6 +753,10 @@ class MediaObject(BaseModel):
 
 
 class Flowstorage(BaseModel):
+    """
+    Gives information on storage for media objects. This schema is for the `http_object_store` media store type which provides URLs for storing media objects in bucket, and is the only store type currently implemented.
+    """
+
     model_config = ConfigDict(
         extra="forbid",
     )
@@ -659,6 +771,10 @@ class Flowstorage(BaseModel):
 
 
 class Service(BaseModel):
+    """
+    Provides information about the store service
+    """
+
     name: Optional[str] = Field(None, description="The service instance name")
     description: Optional[str] = Field(
         None, description="The service instance description"
@@ -685,6 +801,10 @@ class Service(BaseModel):
 
 
 class Collectionitem(BaseModel):
+    """
+    Describes how an object (Source or Flow) is collected into another object of the same type
+    """
+
     id: constr(
         pattern=r"^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
     ) = Field(
@@ -706,6 +826,12 @@ class Flowcollection(RootModel[List[Collectionitem]]):
 
 
 class Source(BaseModel):
+    """
+    Describes a Source: an abstract representation of a piece of media as defined in <https://specs.amwa.tv/ms-04/releases/v1.0.0/docs/2.2._Explanation_-_Source.html>
+
+    Sources may be elemental (and represented directly by a Flow), or may represent a collection of other Sources, e.g. a Source collecting video and audio together.
+    """
+
     id: constr(
         pattern=r"^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
     ) = Field(..., description="Source identifier")
@@ -752,6 +878,10 @@ class Source(BaseModel):
 
 
 class Flowcore(BaseModel):
+    """
+    Describes a Flow (common properties to all Flows, imported by type-specific specifications)
+    """
+
     id: constr(
         pattern=r"^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
     ) = Field(..., description="Flow identifier")
@@ -838,6 +968,10 @@ class Flowcore(BaseModel):
 
 
 class Flowaudio(Flowcore):
+    """
+    Describes an audio Flow
+    """
+
     format: Format1 = Field(
         ..., description="The primary content type URN for the flow."
     )
@@ -849,6 +983,10 @@ class Flowaudio(Flowcore):
 
 
 class Flowdata(Flowcore):
+    """
+    Describes a data Flow
+    """
+
     format: Format2 = Field(
         ..., description="The primary content type URN for the flow."
     )
@@ -860,12 +998,20 @@ class Flowdata(Flowcore):
 
 
 class Flowmulti(Flowcore):
+    """
+    Describes a multi-essence Flow
+    """
+
     format: Format3 = Field(
         ..., description="The primary content type URN for the flow."
     )
 
 
 class Flowvideo(Flowcore):
+    """
+    Describes a video Flow
+    """
+
     format: Format4 = Field(
         ..., description="The primary content type URN for the flow."
     )
