+++ layers/utils/utils.py
@@ -1,56 +1,33 @@
-# pylint: disable=too-many-lines
 import json
 import math
 import os
 import urllib.parse
 import uuid
 from datetime import datetime, timezone
-from enum import Enum
-
-# pylint: disable=no-name-in-module
 from itertools import batched
-from typing import Type
 
 import boto3
 
 # pylint: disable=no-member
 import constants
-import cymple
 from aws_lambda_powertools import Tracer
 from aws_lambda_powertools.utilities.data_classes import APIGatewayProxyEvent
 from aws_lambda_powertools.utilities.data_classes.api_gateway_proxy_event import (
     APIGatewayEventRequestContext,
 )
-from aws_lambda_powertools.utilities.typing import LambdaContext
-from boto3.dynamodb.conditions import And, Attr, Key
 from botocore.config import Config
 from botocore.exceptions import ClientError
-from cymple import QueryBuilder
-from deepdiff import DeepDiff
 from mediatimestamp.immutable import TimeRange
 from params import essence_params, query_params
 from pydantic import BaseModel
-from schema import Flowcollection, Source
 
 tracer = Tracer()
-
 
 events = boto3.client("events")
 sqs = boto3.client("sqs")
 s3 = boto3.client(
     "s3", config=Config(s3={"addressing_style": "virtual"})
 )  # Addressing style is required to ensure pre-signed URLs work as soon as the bucket is created.
-neptune = boto3.client(
-    "neptunedata",
-    region_name=os.environ["AWS_REGION"],
-    endpoint_url=f'https://{os.environ["NEPTUNE_ENDPOINT"]}:8182',
-)
-qb = QueryBuilder()
-
-
-class TimeRangeBoundary(Enum):
-    START = "start"
-    END = "end"
 
 
 @tracer.capture_method(capture_response=False)
@@ -70,161 +47,6 @@ def base_delete_request_dict(
 
 
 @tracer.capture_method(capture_response=False)
-def check_delete_source(source_id: str) -> bool:
-    """Performs a conditional delete on the specified Source. It is only deleted if it is not referenced by any flow representations. Returns True if delete occurred."""
-    query = (
-        qb.match()
-        .node(ref_name="source", labels="source", properties={"id": source_id})
-        .where_literal("NOT exists((source)<-[:represents]-(:flow))")
-        .match()
-        .node(ref_name="source")
-        .related_to(label="has_tags")
-        .node(ref_name="t", labels="tags")
-        .detach_delete(ref_name="source")
-        .return_literal("source.id AS source_id")
-        .get()
-    )
-    query = query.replace(
-        "DETACH DELETE source", "DETACH DELETE source DELETE t"
-    )  # Limitation in Cymple library, unable to stack DELETE
-    results = neptune.execute_open_cypher_query(openCypherQuery=query)
-    return len(results["results"]) > 0
-
-
-@tracer.capture_method(capture_response=False)
-def check_object_exists(bucket, object_id: str) -> bool:
-    """Checks whether the specified object_id (as key) currently exists in the specified S3 Bucket"""
-    try:
-        s3.head_object(Bucket=bucket, Key=object_id)
-        return True
-    except ClientError:
-        return False
-
-
-@tracer.capture_method(capture_response=False)
-def delete_flow_segments(
-    segments_table: "boto3.resources.factory.dynamodb.Table",
-    flow_id: str,
-    parameters: None | dict,
-    timerange_to_delete: TimeRange,
-    context: LambdaContext,
-    s3_queue: str,
-    del_queue: str,
-    item_dict: dict | None = None,
-):
-    """Performs the logic to delete flow segments exits gracefully if within 5 seconds of Lambda timeout"""
-    delete_error = None
-    args = get_key_and_args(flow_id, parameters)
-    args["Limit"] = constants.DELETE_BATCH_SIZE
-    query = segments_table.query(**args)
-    object_ids = set()
-    # Pop first and/or last item in array if they are not entirely covered by the deletion timerange
-    query["Items"] = pop_outliers(timerange_to_delete, query["Items"])
-    if len(query["Items"]) > 0:
-        delete_error = delete_segment_items(
-            segments_table,
-            query["Items"],
-            object_ids,
-        )
-        update_flow_segments_updated(flow_id)
-    # Continue with deletes if no errors, more records available and more than specified milliseconds remain of runtime
-    while (
-        delete_error is None
-        and "LastEvaluatedKey" in query
-        and context.get_remaining_time_in_millis() > constants.LAMBDA_TIME_REMAINING
-    ):
-        query = segments_table.query(
-            **args,
-            ExclusiveStartKey=query["LastEvaluatedKey"],
-        )
-        # Pop first and/or last item in array if they are not entirely covered by the deletion timerange
-        query["Items"] = pop_outliers(timerange_to_delete, query["Items"])
-        if len(query["Items"]) > 0:
-            delete_error = delete_segment_items(
-                segments_table,
-                query["Items"],
-                object_ids,
-            )
-            update_flow_segments_updated(flow_id)
-    # Add affected object_ids to the SQS queue for potential S3 cleanup
-    if len(object_ids) > 0:
-        for message in get_message_batches(list(object_ids)):
-            sqs.send_message(
-                QueueUrl=s3_queue,
-                MessageBody=json.dumps(message),
-            )
-    if item_dict is None:
-        # item_dict only None when called from object_id related segment delete. This method does not support delete requests
-        return
-    # Update DDB record with error if error encountered, no SQS publish to prevent further processing
-    if delete_error:
-        merge_delete_request(
-            {
-                **item_dict,
-                "status": "error",
-                "error": delete_error,
-            }
-        )
-        return
-    # Update DDB record with done, no SQS publish as no further processing required.
-    if "LastEvaluatedKey" not in query:
-        merge_delete_request(
-            {
-                **item_dict,
-                "status": "done",
-                "timerange_remaining": "()",
-            }
-        )
-        return
-    last_timerange = TimeRange.from_str(query["Items"][-1]["timerange"])
-    timerange_remaining = timerange_to_delete.intersect_with(
-        last_timerange.timerange_after()
-    )
-    item_dict["timerange_remaining"] = str(timerange_remaining)
-    item_dict["updated"] = datetime.now().strftime(constants.DATETIME_FORMAT)
-    put_deletion_request(del_queue, item_dict)
-
-
-@tracer.capture_method(capture_response=False)
-def delete_segment_items(
-    segments_table: "boto3.resources.factory.dynamodb.Table",
-    items: list[dict],
-    object_ids: set[str],
-) -> dict | None:
-    """Loop supplied items and delete, early return on error, append to object_ids supplied on success"""
-    delete_error = None
-    for item in items:
-        key = {
-            "flow_id": item["flow_id"],
-            "timerange_end": item["timerange_end"],
-        }
-        try:
-            delete_item = segments_table.delete_item(
-                Key=key,
-                ReturnValues="ALL_OLD",
-            )
-            if "Attributes" in delete_item:
-                object_ids.add(item["object_id"])
-                publish_event(
-                    "flows/segments_deleted",
-                    {"flow_id": item["flow_id"], "timerange": item["timerange"]},
-                    [f'tams:flow:{item["flow_id"]}'],
-                )
-        except ClientError as e:
-            delete_error = {
-                "type": e.response["Error"]["Code"],
-                "summary": e.response["Error"]["Message"],
-                "traceback": [
-                    f"Delete Segment Key: {json.dumps(key, default=str)}",
-                    json.dumps(e.response["ResponseMetadata"], default=str),
-                ],
-                "time": datetime.now().strftime("%Y-%m-%dT%H:%M:%SZ"),
-            }
-            break
-    return delete_error
-
-
-@tracer.capture_method(capture_response=False)
 def generate_link_url(current_event: APIGatewayProxyEvent, page_value: str) -> str:
     """Generates a link URL relative to the API Gateway request that calls it"""
     host = current_event.request_context.domain_name
@@ -239,138 +61,7 @@ def generate_link_url(current_event: APIGatewayProxyEvent, page_value: str) -> s
         if current_event.query_string_parameters
         else ""
     )
-
     return f'<https://{host}{path}?{query_string}page={urllib.parse.quote_plus(page_value)}>; rel="next"'
-
-
-@tracer.capture_method(capture_response=False)
-# pylint: disable=W0102:dangerous-default-value
-def generate_presigned_url(
-    method: str, bucket: str, key: str, other_args: dict = {}
-) -> str:
-    """Generates an S3 pre-signed URL"""
-    url = s3.generate_presigned_url(
-        ClientMethod=method,
-        Params={
-            "Bucket": bucket,
-            "Key": key,
-            **other_args,
-        },
-        ExpiresIn=3600,
-    )
-    return url
-
-
-@tracer.capture_method(capture_response=False)
-def get_flow_timerange(
-    table: "boto3.resources.factory.dynamodb.Table", flow_id: str
-) -> str:
-    """Get the timerange for a specified flow"""
-    first_segment = table.query(
-        KeyConditionExpression=Key("flow_id").eq(flow_id),
-        Limit=1,
-        ScanIndexForward=True,
-        Select="SPECIFIC_ATTRIBUTES",
-        ProjectionExpression="timerange",
-    )["Items"]
-    last_segment = table.query(
-        KeyConditionExpression=Key("flow_id").eq(flow_id),
-        Limit=1,
-        ScanIndexForward=False,
-        Select="SPECIFIC_ATTRIBUTES",
-        ProjectionExpression="timerange",
-    )["Items"]
-    if len(first_segment) > 0 and len(last_segment) > 0:
-        return str(
-            TimeRange.from_str(
-                first_segment[0]["timerange"]
-            ).extend_to_encompass_timerange(
-                TimeRange.from_str(last_segment[0]["timerange"])
-            )
-        )
-    if len(first_segment) > 0:
-        return first_segment[0]["timerange"]
-    if len(last_segment) > 0:
-        return last_segment[0]["timerange"]
-    return "()"
-
-
-@tracer.capture_method(capture_response=False)
-def get_timerange_expression(
-    expression_type: Type[Attr] | Type[Key],
-    boundary: TimeRangeBoundary,
-    filter_value: TimeRange,
-):
-    """Returns a DynamoDB expression, for Key or Filter, to add the specfied timerange condition."""
-    other_boundary = (
-        TimeRangeBoundary.START
-        if boundary == TimeRangeBoundary.END
-        else TimeRangeBoundary.END
-    )
-    include_method = getattr(filter_value, f"includes_{other_boundary.value}")
-    nanosec_value = getattr(filter_value, other_boundary.value).to_nanosec()
-    match boundary:
-        case TimeRangeBoundary.START:
-            if include_method():
-                return expression_type(f"timerange_{boundary.value}").lte(nanosec_value)
-            else:
-                return expression_type(f"timerange_{boundary.value}").lt(nanosec_value)
-        case TimeRangeBoundary.END:
-            if include_method():
-                return expression_type(f"timerange_{boundary.value}").gte(nanosec_value)
-            else:
-                return expression_type(f"timerange_{boundary.value}").gt(nanosec_value)
-
-
-@tracer.capture_method(capture_response=False)
-def get_key_and_args(flow_id: str, parameters: None | dict) -> dict | bool:
-    """Generate key expression and args for a dynamodb query operation"""
-    args = {
-        "KeyConditionExpression": Key("flow_id").eq(flow_id),
-        "ScanIndexForward": True,
-        "Limit": constants.DEFAULT_PAGE_LIMIT,
-    }
-    if parameters is None:
-        return args
-    reverse_order = parameters.get("reverse_order", "false").lower() == "true"
-    args["ScanIndexForward"] = not reverse_order
-    # Pagination query string parameters
-    if "limit" in parameters:
-        args["Limit"] = min(int(parameters["limit"]), constants.MAX_PAGE_LIMIT)
-    if "page" in parameters:
-        args["ExclusiveStartKey"] = {
-            "flow_id": flow_id,
-            "timerange_end": int(parameters["page"]),
-        }
-    # Parse timerange filter out of parameters
-    timerange_filter = (
-        TimeRange.from_str(parameters["timerange"])
-        if "timerange" in parameters
-        else None
-    )
-    # Update Key Expression
-    if "object_id" in parameters:
-        args["IndexName"] = "object-id-index"
-        args["KeyConditionExpression"] = And(
-            args["KeyConditionExpression"],
-            Key("object_id").eq(parameters["object_id"]),
-        )
-    elif timerange_filter and timerange_filter.start:
-        args["KeyConditionExpression"] = And(
-            args["KeyConditionExpression"],
-            get_timerange_expression(Key, TimeRangeBoundary.END, timerange_filter),
-        )
-    # Build Filter expression
-    if timerange_filter:
-        if timerange_filter.start and "object_id" in parameters:
-            args["FilterExpression"] = get_timerange_expression(
-                Attr, TimeRangeBoundary.END, timerange_filter
-            )
-        if timerange_filter.end:
-            args["FilterExpression"] = get_timerange_expression(
-                Attr, TimeRangeBoundary.START, timerange_filter
-            )
-    return args
 
 
 @tracer.capture_method(capture_response=False)
@@ -383,6 +74,16 @@ def get_message_batches(items: list) -> list:
     )
     batch_size = math.ceil(len(items) / batch_count)
     return list(batched(items, batch_size))
+
+
+@tracer.capture_method(capture_response=False)
+def put_message_batches(queue: str, items: list) -> list:
+    """Batch sends a list of message items to the specified queue url"""
+    for message in get_message_batches(items):
+        sqs.send_message(
+            QueueUrl=queue,
+            MessageBody=json.dumps(message),
+        )
 
 
 @tracer.capture_method(capture_response=False)
@@ -440,121 +141,6 @@ def pop_outliers(timerange: TimeRange, items: list) -> list:
 
 
 @tracer.capture_method(capture_response=False)
-def get_flow_source_id(flow_id: str) -> str | None:
-    """Get the source_id for the specified Flow"""
-    try:
-        query = (
-            qb.match()
-            .node(labels="flow", properties={"id": flow_id})
-            .related_to(label="represents")
-            .node(ref_name="s", labels="source")
-            .return_literal("s.id as source_id")
-            .get()
-        )
-        results = neptune.execute_open_cypher_query(openCypherQuery=query)
-        return results["results"][0]["source_id"]
-    except IndexError:
-        return None
-
-
-@tracer.capture_method(capture_response=False)
-def get_source_collected_by(source_id: str) -> list:
-    """Get the collect_by source ids for the specified Source"""
-    try:
-        query = (
-            qb.match()
-            .node(labels="source", properties={"id": source_id})
-            .related_from(label="represents")
-            .node(labels="flow")
-            .related_to(label="collected_by")
-            .node(labels="flow")
-            .related_to(label="represents")
-            .node(ref_name="s", labels="source")
-            .return_literal("collect(s.id) as source_collected_by")
-            .get()
-        )
-        results = neptune.execute_open_cypher_query(openCypherQuery=query)
-        return results["results"][0]["source_collected_by"]
-    except IndexError:
-        return []
-
-
-@tracer.capture_method(capture_response=False)
-def get_flow_collected_by(flow_id: str) -> list:
-    """Get the collect_by flow ids for the specified Flow"""
-    try:
-        query = (
-            qb.match()
-            .node(labels="flow", properties={"id": flow_id})
-            .related_to(label="collected_by")
-            .node(ref_name="f", labels="flow")
-            .return_literal("collect(f.id) as flow_collected_by")
-            .get()
-        )
-        results = neptune.execute_open_cypher_query(openCypherQuery=query)
-        return results["results"][0]["flow_collected_by"]
-    except IndexError:
-        return []
-
-
-@tracer.capture_method(capture_response=False)
-# pylint: disable=dangerous-default-value
-def publish_event(detail_type: str, details: dict, resources: list = []) -> None:
-    """Publishes the supplied events to an EventBridge EventBus"""
-
-    if all(not r.startswith("tams:source:") for r in resources) and any(
-        r.startswith("tams:flow:") for r in resources
-    ):
-        flow_id = next(
-            r[len("tams:flow:") :] for r in resources if r.startswith("tams:flow:")
-        )
-        source_id = get_flow_source_id(flow_id)
-        if source_id:
-            resources.append(f"tams:source:{source_id}")
-    if all(not r.startswith("tams:source-collected-by:") for r in resources) and any(
-        r.startswith("tams:source:") for r in resources
-    ):
-        source_id = next(
-            r[len("tams:source:") :] for r in resources if r.startswith("tams:source:")
-        )
-        resources.extend(
-            f"tams:source-collected-by:{s_id}"
-            for s_id in get_source_collected_by(source_id)
-        )
-    if all(not r.startswith("tams:flow-collected-by:") for r in resources) and any(
-        r.startswith("tams:flow:") for r in resources
-    ):
-        flow_id = next(
-            r[len("tams:flow:") :] for r in resources if r.startswith("tams:flow:")
-        )
-        resources.extend(
-            f"tams:flow-collected-by:{s_id}" for s_id in get_flow_collected_by(flow_id)
-        )
-    events.put_events(
-        Entries=[
-            {
-                "Source": "tams.api",
-                "EventBusName": os.environ["EVENT_BUS"],
-                "DetailType": detail_type,
-                "Time": datetime.now(),
-                "Detail": json.dumps(details),
-                "Resources": resources,
-            }
-        ],
-    )
-
-
-@tracer.capture_method(capture_response=False)
-def put_deletion_request(queue: str, item: dict) -> None:
-    """Publishs a message to SQS and inserts into dynamodb"""
-    sqs.send_message(
-        QueueUrl=queue,
-        MessageBody=json.dumps(item),
-    )
-    merge_delete_request(item)
-
-
-@tracer.capture_method(capture_response=False)
 def remove_null(obj: dict | list) -> None:
     """Removes null and other "empty" keys from a dict/list recursively"""
     if isinstance(obj, list):
@@ -575,38 +161,6 @@ def remove_null(obj: dict | list) -> None:
                         pass
             else:
                 remove_null(v)
-
-
-@tracer.capture_method(capture_response=False)
-def update_flow_segments_updated(flow_id: str) -> None:
-    """Update the segments_updated field on the specified Flow"""
-    try:
-        item_dict = set_node_property_base(
-            "flow",
-            flow_id,
-            {
-                "flow.segments_updated": datetime.now()
-                .astimezone(timezone.utc)
-                .strftime(constants.DATETIME_FORMAT)
-            },
-        )
-        publish_event(
-            "flows/updated",
-            {"flow": item_dict},
-            [
-                f'tams:flow:{item_dict["id"]}',
-                f'tams:source:{item_dict["source_id"]}',
-                *[
-                    f"tams:flow-collected-by:{c_id}"
-                    for c_id in item_dict.get("collected_by", [])
-                ],
-            ],
-        )
-    except ValueError:
-        # The set_node_property_base function will throw an exception
-        # if specified flow does not exist. When setting the segments_updated
-        # field in the database don't need to worry if the flow does not exist.
-        return
 
 
 @tracer.capture_method(capture_response=False)
@@ -664,218 +218,6 @@ def deserialise_neptune_obj(obj: dict) -> dict:
 
 
 @tracer.capture_method(capture_response=False)
-# pylint: disable=dangerous-default-value
-def generate_source_query(
-    properties: dict, set_dict: dict = None, where_literals: list = []
-) -> cymple.builder.NodeAvailable:
-    """Returns an Open Cypher Match query to return specified Source"""
-    query = (
-        qb.match()
-        .node(ref_name="f", labels="flow", properties=properties.get("flow", {}))
-        .related_to(label="represents")
-        .node(
-            ref_name="source",
-            labels="source",
-            properties=properties.get("source", {}),
-        )
-        .related_to(label="has_tags")
-        .node(ref_name="t", labels="tags", properties=properties.get("tags", {}))
-    )
-    if set_dict:
-        query = query.set(set_dict).with_("*")
-    if len(where_literals) > 0:
-        query = query.where_literal(" AND ".join(where_literals))
-    query = (
-        query.match_optional()
-        .node(ref_name="f")
-        .related_from(ref_name="c", label="collected_by")
-        .node(labels="flow")
-        .related_to(label="represents")
-        .node(ref_name="sc", labels="source")
-        .match_optional()
-        .node(ref_name="f")
-        .related_to(label="collected_by")
-        .node(labels="flow")
-        .related_to(label="represents")
-        .node(ref_name="cb", labels="source")
-    )
-    return query
-
-
-@tracer.capture_method(capture_response=False)
-# pylint: disable=dangerous-default-value
-def generate_flow_query(
-    properties: dict, set_dict: dict = None, where_literals: list = []
-) -> cymple.builder.NodeAvailable:
-    """Returns an Open Cypher Match query to return specified Flow"""
-    query = (
-        qb.match()
-        .node(
-            ref_name="flow",
-            labels="flow",
-            properties=properties.get("flow", {}),
-        )
-        .related_to(label="represents")
-        .node(
-            ref_name="s",
-            labels="source",
-            properties=properties.get("source", {}),
-        )
-        .match()
-        .node(ref_name="flow")
-        .related_to(label="has_essence_parameters")
-        .node(
-            ref_name="e",
-            labels="essence_parameters",
-            properties=properties.get("essence_parameters", {}),
-        )
-        .match()
-        .node(ref_name="flow")
-        .related_to(label="has_tags")
-        .node(ref_name="t", labels="tags", properties=properties.get("tags", {}))
-    )
-    if set_dict:
-        query = query.set(set_dict).with_("*")
-    if len(where_literals) > 0:
-        query = query.where_literal(" AND ".join(where_literals))
-    query = (
-        query.match_optional()
-        .node(ref_name="flow")
-        .related_from(ref_name="c", label="collected_by")
-        .node(ref_name="fc", labels="flow")
-        .match_optional()
-        .node(ref_name="flow")
-        .related_to(label="collected_by")
-        .node(ref_name="cb", labels="flow")
-    )
-    return query
-
-
-@tracer.capture_method(capture_response=False)
-# pylint: disable=dangerous-default-value
-def generate_delete_request_query(
-    properties: dict, set_dict: dict = None, where_literals: list = []
-) -> cymple.builder.NodeAvailable:
-    """Returns an Open Cypher Match query to return specified Delete Request"""
-    query = (
-        qb.match()
-        .node(
-            ref_name="delete_request",
-            labels="delete_request",
-            properties=properties.get("delete_request", {}),
-        )
-        .related_to(label="has_error")
-        .node(ref_name="e", labels="error", properties=properties.get("error", {}))
-    )
-    if set_dict:
-        query = query.set(set_dict).with_("*")
-    if len(where_literals) > 0:
-        query = query.where_literal(" AND ".join(where_literals))
-    return query
-
-
-@tracer.capture_method(capture_response=False)
-# pylint: disable=dangerous-default-value
-def generate_match_query(
-    record_type: str, properties: dict, set_dict: dict = None, where_literals: list = []
-) -> cymple.builder.NodeAvailable:
-    """Returns an Open Cypher Match query to return specified record type"""
-    match record_type:
-        case "source":
-            return generate_source_query(properties, set_dict, where_literals)
-        case "flow":
-            return generate_flow_query(properties, set_dict, where_literals)
-        case "delete_request":
-            return generate_delete_request_query(properties, set_dict, where_literals)
-
-
-@tracer.capture_method(capture_response=False)
-def check_node_exists(record_type: str, record_id: str) -> bool:
-    """Checks whether the specified Node exists in the Neptune Database"""
-    query = (
-        qb.match()
-        .node(ref_name="n", labels=record_type, properties={"id": record_id})
-        .return_literal("n.id")
-        .get()
-    )
-    results = neptune.execute_open_cypher_query(openCypherQuery=query)
-    return len(results["results"]) > 0
-
-
-@tracer.capture_method(capture_response=False)
-def query_node_tags(record_type: str, record_id: str) -> dict:
-    """Returns the TAMS Tags for the specified Node"""
-    try:
-        query = (
-            qb.match()
-            .node(labels=record_type, properties={"id": record_id})
-            .related_to(label="has_tags")
-            .node(ref_name="t", labels="tags")
-            .return_literal("t {.*} AS tags")
-            .get()
-        )
-        results = neptune.execute_open_cypher_query(openCypherQuery=query)
-        return results["results"][0]["tags"]
-    except IndexError as e:
-        raise ValueError("No results returned from the database query.") from e
-
-
-@tracer.capture_method(capture_response=False)
-def query_node_property(record_type: str, record_id: str, prop_name: str) -> any:
-    """Returns the value of the specified Node property"""
-    try:
-        query = (
-            qb.match()
-            .node(ref_name="n", labels=record_type, properties={"id": record_id})
-            .return_literal(f"n.{prop_name} AS property")
-            .get()
-        )
-        results = neptune.execute_open_cypher_query(openCypherQuery=query)
-        return results["results"][0]["property"]
-    except IndexError as e:
-        raise ValueError("No results returned from the database query.") from e
-
-
-@tracer.capture_method(capture_response=False)
-def query_flow_collection(flow_id: str) -> list:
-    """Returns the flow_collection of the specified Flow"""
-    try:
-        query = (
-            qb.match()
-            .node(ref_name="f", labels="flow", properties={"id": flow_id})
-            .match_optional()
-            .node(ref_name="f")
-            .related_from(ref_name="c", label="collected_by")
-            .node(ref_name="fc", labels="flow")
-            .return_literal("f.id as id, collect(c {.*, id: fc.id}) AS flow_collection")
-            .get()
-        )
-        results = neptune.execute_open_cypher_query(openCypherQuery=query)
-        return results["results"][0]["flow_collection"]
-    except IndexError as e:
-        raise ValueError("No results returned from the database query.") from e
-
-
-@tracer.capture_method(capture_response=False)
-def query_node(record_type: str, record_id: str) -> dict:
-    """Returns the specified Node from the Neptune Database"""
-    try:
-        query = (
-            generate_match_query(record_type, {record_type: {"id": record_id}})
-            .return_literal(constants.RETURN_LITERAL[record_type])
-            .get()
-        )
-        results = neptune.execute_open_cypher_query(openCypherQuery=query)
-        deserialised_results = [
-            deserialise_neptune_obj(result[record_type])
-            for result in results["results"]
-        ]
-        return deserialised_results[0]
-    except IndexError as e:
-        raise ValueError("No results returned from the database query.") from e
-
-
-@tracer.capture_method(capture_response=False)
 def parse_parameters(parameters: dict) -> tuple[int, int, dict, list]:
     """Parses API Gateway parameters into the structure used by OpenCypher query"""
     source_id = parameters.get("source_id", None)
@@ -916,366 +258,59 @@ def parse_parameters(parameters: dict) -> tuple[int, int, dict, list]:
 
 
 @tracer.capture_method(capture_response=False)
-def query_sources(parameters: dict) -> tuple[list, int]:
-    """Returns a list of the TAMS Sources from the Neptune Database"""
-    page, limit, props, where_literals = parse_parameters(parameters)
-    query = generate_source_query(
-        {
-            "source": props["properties"],
-            "tags": props["tag_properties"],
-        },
-        where_literals=where_literals,
-    )
-    query = (
-        query.return_literal(constants.RETURN_LITERAL["source"])
-        .order_by("source.id")
-        .skip(page)
-        .limit(limit)
-        .get()
-    )
-    results = neptune.execute_open_cypher_query(openCypherQuery=query)
-    deserialised_results = [
-        deserialise_neptune_obj(result["source"]) for result in results["results"]
-    ]
-    next_page = page + limit if len(deserialised_results) == limit else None
-    return deserialised_results, next_page
-
-
-@tracer.capture_method(capture_response=False)
-def query_flows(parameters: dict) -> tuple[list, int]:
-    """Returns a list of the TAMS Flows from the Neptune Database"""
-    page, limit, props, where_literals = parse_parameters(parameters)
-    query = generate_flow_query(
-        {
-            "flow": props["properties"],
-            "source": props["source_properties"],
-            "essence_parameters": props["essence_properties"],
-            "tags": props["tag_properties"],
-        },
-        where_literals=where_literals,
-    )
-    query = (
-        query.return_literal(constants.RETURN_LITERAL["flow"])
-        .order_by("flow.id")
-        .skip(page)
-        .limit(limit)
-        .get()
-    )
-    results = neptune.execute_open_cypher_query(openCypherQuery=query)
-    deserialised_results = [
-        deserialise_neptune_obj(result["flow"]) for result in results["results"]
-    ]
-    next_page = page + limit if len(deserialised_results) == limit else None
-    return deserialised_results, next_page
-
-
-@tracer.capture_method(capture_response=False)
-def query_delete_requests() -> list:
-    """Returns a list of the TAMS Delete Request from the Neptune Database"""
-    query = generate_delete_request_query({})
-    query = (
-        query.return_literal(constants.RETURN_LITERAL["delete_request"])
-        .order_by("delete_request.id")
-        .get()
-    )
-    results = neptune.execute_open_cypher_query(openCypherQuery=query)
-    deserialised_results = [
-        deserialise_neptune_obj(result["delete_request"])
-        for result in results["results"]
-    ]
-    return deserialised_results
-
-
-@tracer.capture_method(capture_response=False)
 def filter_dict(obj: dict, keys: set) -> dict:
     """Returns a dictionary with specific keys removed"""
     return {k: v for k, v in obj.items() if k not in keys}
 
 
 @tracer.capture_method(capture_response=False)
-def validate_flow_collection(flow_id: str, flow_collection: Flowcollection):
-    """Checks whether the supplied Flow Collection is valid"""
-    if not flow_collection:
+def publish_event(detail_type: str, details: dict, resources) -> None:
+    """Publishes the supplied events to an EventBridge EventBus"""
+    events.put_events(
+        Entries=[
+            {
+                "Source": "tams.api",
+                "EventBusName": os.environ["EVENT_BUS"],
+                "DetailType": detail_type,
+                "Time": datetime.now(),
+                "Detail": json.dumps(details),
+                "Resources": resources,
+            }
+        ],
+    )
+
+
+@tracer.capture_method(capture_response=False)
+def put_message(queue: str, item: dict) -> None:
+    """Publishs a message to SQS"""
+    sqs.send_message(
+        QueueUrl=queue,
+        MessageBody=json.dumps(item),
+    )
+
+
+@tracer.capture_method(capture_response=False)
+def check_object_exists(bucket, object_id: str) -> bool:
+    """Checks whether the specified object_id (as key) currently exists in the specified S3 Bucket"""
+    try:
+        s3.head_object(Bucket=bucket, Key=object_id)
         return True
-    for collection in flow_collection.root:
-        if flow_id == collection.id or not check_node_exists("flow", collection.id):
-            return False
-    return True
+    except ClientError:
+        return False
 
 
 @tracer.capture_method(capture_response=False)
-def merge_source(source_dict: dict) -> None:
-    """Perform an OpenCypher Merge operation on the supplied TAMS Source record"""
-    tags = source_dict.get("tags", {})
-    query = (
-        qb.merge()
-        .node(
-            ref_name="s",
-            labels="source",
-            properties={"id": source_dict["id"]},
-        )
-        .related_to(label="has_tags")
-        .node(ref_name="t", labels="tags")
-        .set(serialise_neptune_obj(filter_dict(source_dict, {"id", "tags"}), "s."))
-    )
-    # Add Set for Source Tags
-    if tags:
-        query = query.set(serialise_neptune_obj(tags, "t."))
-    neptune.execute_open_cypher_query(openCypherQuery=query.get())
-
-
-@tracer.capture_method(capture_response=False)
-def merge_flow(flow_dict: dict, existing_dict: dict) -> dict:
-    """Perform an OpenCypher Merge operation on the supplied TAMS Flow record"""
-    # Extract properties required for other node types
-    tags = flow_dict.get("tags", {})
-    essence_parameters = flow_dict.get("essence_parameters", {})
-    flow_collection = flow_dict.get("flow_collection", [])
-    flow_properties = filter_dict(
-        flow_dict, {"id", "source_id", "tags", "essence_parameters", "flow_collection"}
-    )
-    if existing_dict:
-        existing_flow_collection = existing_dict.get("flow_collection", [])
-        # If there is an flow collection and it has changed delete it so that it is set correctly with the merge
-        if existing_flow_collection and DeepDiff(
-            flow_collection, existing_flow_collection, ignore_order=True
-        ):
-            set_flow_collection(flow_dict["id"], "temp", [])
-        null_tags = {
-            k: None for k in existing_dict.get("tags", {}).keys() - tags.keys()
-        }
-        null_essence_parameters = {
-            k: None
-            for k in existing_dict.get("essence_parameters", {}).keys()
-            - essence_parameters.keys()
-        }
-        null_properties = {
-            k: None
-            for k in (
-                existing_dict.keys()
-                - {
-                    "id",
-                    "source_id",
-                    "tags",
-                    "essence_parameters",
-                    "flow_collection",
-                    "created_by",
-                    "updated_by",
-                }
-            )
-            - flow_properties.keys()
-        }
-        tags = tags | null_tags
-        essence_parameters = essence_parameters | null_essence_parameters
-        flow_properties = flow_properties | null_properties
-    # Build Merge queries
-    query = (
-        qb.match()
-        .node(ref_name="s", labels="source", properties={"id": flow_dict["source_id"]})
-        .merge()
-        .node(ref_name="f", labels="flow", properties={"id": flow_dict["id"]})
-        .related_to(label="represents")
-        .node(ref_name="s")
-        .merge()
-        .node(ref_name="f")
-        .related_to(label="has_tags")
-        .node(ref_name="t", labels="tags")
-        .merge()
-        .node(ref_name="f")
-        .related_to(label="has_essence_parameters")
-        .node(ref_name="ep", labels="essence_parameters")
-        .set(serialise_neptune_obj(flow_properties, "f."))
-    )
-    # Add Set for Flow Tags
-    if tags:
-        query = query.set(serialise_neptune_obj(tags, "t."))
-    # Add Set for Flow Essence Parameters
-    if essence_parameters:
-        query = query.set(serialise_neptune_obj(essence_parameters, "ep."))
-    # Add flow collection queries as needed
-    query_collection = generate_flow_collection_query(flow_collection)
-    if query_collection:
-        query = query + query_collection
-    neptune.execute_open_cypher_query(openCypherQuery=query.get())
-    # Too complex to try and get OpenCypher to return the object in the same query so calling the DB to get it separately
-    return query_node("flow", flow_dict["id"])
-
-
-@tracer.capture_method(capture_response=False)
-def merge_source_flow(flow_dict: dict, existing_dict: dict) -> dict:
-    """Perform an OpenCypher Merge operation on the supplied TAMS Source/Flow record"""
-    # Check if supplied source already exists, create if not
-    if not check_node_exists("source", flow_dict["source_id"]):
-        source: Source = Source(**flow_dict)
-        source.id = flow_dict["source_id"]
-        source_dict = model_dump(source)
-        merge_source(source_dict)
-        publish_event(
-            "sources/created",
-            {"source": source_dict},
-            [f'tams:source:{source_dict["id"]}'],
-        )
-    # Create Flow
-    return merge_flow(flow_dict, existing_dict)
-
-
-@tracer.capture_method(capture_response=False)
-def merge_delete_request(delete_request_dict: dict) -> None:
-    """Perform an OpenCypher Merge operation on the supplied TAMS Delete Request record"""
-    error = delete_request_dict.get("error", {})
-    query = (
-        qb.merge()
-        .node(
-            ref_name="d",
-            labels="delete_request",
-            properties={"id": delete_request_dict["id"]},
-        )
-        .related_to(label="has_error")
-        .node(ref_name="e", labels="error")
-        .set(
-            serialise_neptune_obj(
-                filter_dict(delete_request_dict, {"id", "error"}), "d."
-            )
-        )
-    )
-    # Add Set for Error
-    if error:
-        query = query.set(serialise_neptune_obj(error, "e."))
-    neptune.execute_open_cypher_query(openCypherQuery=query.get())
-
-
-@tracer.capture_method(capture_response=False)
-def delete_flow(flow_id: str) -> str | None:
-    """Deletes the specified Flow from the Neptune Database"""
-    try:
-        query = (
-            qb.match()
-            .node(ref_name="flow", labels="flow", properties={"id": flow_id})
-            .related_to(label="represents")
-            .node(ref_name="s", labels="source")
-            .match()
-            .node(ref_name="flow")
-            .related_to(label="has_tags")
-            .node(ref_name="t", labels="tags")
-            .match()
-            .node(ref_name="flow")
-            .related_to(label="has_essence_parameters")
-            .node(ref_name="e", labels="essence_parameters")
-            .detach_delete(ref_name="flow")
-            .return_literal("s.id AS source_id")
-            .get()
-        )
-        query = query.replace(
-            "DETACH DELETE flow", "DETACH DELETE flow DELETE t DELETE e"
-        )  # Limitation in Cymple library, unable to stack DELETE
-        results = neptune.execute_open_cypher_query(openCypherQuery=query)
-        return results["results"][0]["source_id"]
-    except IndexError:
-        return None
-
-
-@tracer.capture_method(capture_response=False)
-def set_node_property_base(record_type: str, record_id: str, props: dict) -> dict:
-    """Performs an OpenCypher Set operation on the specified Node and properties"""
-    try:
-        query = (
-            generate_match_query(
-                record_type,
-                {record_type: {"id": record_id}},
-                set_dict=props,
-            )
-            .return_literal(constants.RETURN_LITERAL[record_type])
-            .get()
-        )
-        results = neptune.execute_open_cypher_query(openCypherQuery=query)
-        deserialised_results = [
-            deserialise_neptune_obj(result[record_type])
-            for result in results["results"]
-        ]
-        return deserialised_results[0]
-    except IndexError as e:
-        raise ValueError("No results returned from the database query.") from e
-
-
-@tracer.capture_method(capture_response=False)
-def set_node_property(
-    record_type: str, record_id: str, username: str, props: dict
-) -> dict:
-    """Performs an OpenCypher Set operation on the specified Node and properties with the addition of updated and updated_by properties"""
-    meta_props = {
-        **props,
-        f"{record_type}.{"metadata_" if record_type == "flow" else ""}updated": datetime.now()
-        .astimezone(timezone.utc)
-        .strftime(constants.DATETIME_FORMAT),
-        f"{record_type}.updated_by": username,
-    }
-    return set_node_property_base(record_type, record_id, meta_props)
-
-
-@tracer.capture_method(capture_response=False)
-def generate_flow_collection_query(
-    flow_collection: list, set_dict: dict | None = None
-) -> cymple.builder.SetAvailable | None:
-    """Returns a QueryBuilder that creates the specfied flow collection."""
-    if not set_dict:
-        set_dict = {}
-    if not flow_collection:
-        return None
-    query = qb.with_("f")
-    # process the supplied flow_collection into numbered records
-    ref_names = []
-    for n, collection in enumerate(flow_collection):
-        collection_properties = {
-            k: json.dumps(v) if isinstance(v, list) or isinstance(v, dict) else v
-            for k, v in collection.items()
-            if k != "id"
-        }
-        ref_names.append((f"f{n}", f"c{n}", collection["id"], collection_properties))
-    # Add the match queries for each flow_collection record
-    for f_ref, _, f_id, _ in ref_names:
-        query = query.match().node(
-            ref_name=f_ref, labels="flow", properties={"id": f_id}
-        )
-    # Add the merge queries to create the collected_by edges for each flow_collection record
-    for f_ref, c_ref, _, _ in ref_names:
-        query = (
-            query.merge()
-            .node(ref_name="f")
-            .related_from(ref_name=c_ref, label="collected_by")
-            .node(ref_name=f_ref)
-        )
-    # Build the dict of set operations to carry out
-    for _, c_ref, _, props in ref_names:
-        for k, v in props.items():
-            set_dict[f"{c_ref}.{k}"] = v
-    return query.set(set_dict)
-
-
-@tracer.capture_method(capture_response=False)
-def set_flow_collection(flow_id: str, username: str, flow_collection: list) -> dict:
-    """Set the specfified flow to have the specified flow_collection and set flow metadata_updated and updated_by properties"""
-    # Create initial query that deletes all collect_by edges for this flow
-    query = (
-        qb.match()
-        .node(ref_name="f", labels="flow", properties={"id": flow_id})
-        .match_optional()
-        .node(ref_name="f")
-        .related_from(ref_name="c", label="collected_by")
-        .node(labels="flow")
-        .delete(ref_name="c")
-    )
-    # Create the base set query
-    query_collection = generate_flow_collection_query(
-        flow_collection,
-        {
-            "f.metadata_updated": datetime.now()
-            .astimezone(timezone.utc)
-            .strftime(constants.DATETIME_FORMAT),
-            "f.updated_by": username,
+def generate_presigned_url(
+    method: str, bucket: str, key: str, **kwargs: None | dict
+) -> str:
+    """Generates an S3 pre-signed URL"""
+    url = s3.generate_presigned_url(
+        ClientMethod=method,
+        Params={
+            "Bucket": bucket,
+            "Key": key,
+            **kwargs,
         },
+        ExpiresIn=3600,
     )
-    if query_collection:
-        query = query + query_collection
-    neptune.execute_open_cypher_query(openCypherQuery=query.get())
-    # Too complex to try and get OpenCypher to return the object in the same query so calling the DB to get it separately
-    return query_node("flow", flow_id)
+    return url
